+++ database/migrations/2026_05_06_201917_create_biz_stage_table.php
@@ -59,10 +59,10 @@ return new class extends Migration
             // 检查是否存在数据
             $exists = DB::table($this->name)->exists();
             // 不存在数据时，删除表
-            // if (! $exists) {
-            // 删除 reverse
-            Schema::dropIfExists($this->name);
-            // }
+            if (! $exists) {
+                // 删除 reverse
+                Schema::dropIfExists($this->name);
+            }
         }
     }
 };
